1.52 TZ11 - add LikeDbStorage - update FilmDbStorage

+++ src/main/java/ru/yandex/practicum/filmorate/model/FilmGenre.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class FilmGenre {
-    private Integer filmId;
+    private Long filmId;
     private Integer genreId;
 
-    public FilmGenre(Integer filmId, Integer genreId) {
+    public FilmGenre(Long filmId, Integer genreId) {
         this.filmId = filmId;
         this.genreId = genreId;
     }
